+++ main.py
@@ -33,7 +33,7 @@ from .pixivSDK import pixiv, refresh_pixiv_host_map
 from .utils import command_usage, help_text
 
 
-@register("pixivdirect", "Sagiri777", "PixivDirect command plugin", "2.0.0")
+@register("pixivdirect", "Sagiri777", "PixivDirect command plugin", "2.0.1")
 class PixivDirectPlugin(Star):
     def __init__(self, context: Context):
         super().__init__(context)